Use PHP Markdown Extra's Definition Lists where applicable The definition list is best suited to selected lists. No bullet needed and it only distracts attention.
Following the PHP Markdown Extra,
> Terms must be separated from the previous definition by a blank line.
(it is applied in this commit).

The appearance of the term (dt) is already bold, so it has been removed
from the MD sources.

Only the appearance of the description (dd) requires correction on the site.
Please restore the left indent with something like:

```css
dl dd {
    margin-left: 2em;
}
```