Use reference style for all links, add rel for external links Best practices is to use `rel="noopener noreferrer"`, but we can't use spaces in the attribute value using PHP Markdown Extra.
Notes and warnings begin and end was marked with html comments.
Some notes can span few paragraphs and code inside it, so notes and
warnings should be block level elements.
TODO: PHP Markdown Extra Definition List is the best solution for notes and
warnings:
8<--------
Note:
: para1

  ```
  code
  ```

  para2
8<--------
But we need to wrap it to a div like this:
<div markdown="1" class="note"> ... </div>
To style this snippet isn't a problem...