temporarily remove SF3 support
This is the minimum amount of changes necessary to make this still
build and work with sources/lib/sf3/ removed. All in one commit so
it can later (when the licence incompatibility will be resolved)
easily be “git revert”ed.

This is required to be able to legally distribute Polyphone binaries.