updated sf3 (sftools) source license to LGPLv2 after MuseScore upstream change
MuseScore's sftools (http://github.com/MuseScore/sftools)
was the base for the source code in lib/sf3. This code was
formerly licensed under GPLv2-only, which is incompatible
with polyphone's GPLv3 license. MuseScore has now updated
the license for this code to LGPLv2, which allows the
corresponding code in polyphone to also be placed under this
license.